+++ lib/simdart.dart
@@ -2,7 +2,7 @@ export 'src/event.dart';
 export 'src/sim_context.dart';
 export 'src/interval.dart';
 export 'src/observable.dart';
-export 'src/resource_configurator.dart' hide ResourcesConfiguratorHelper;
+export 'src/resources.dart' hide ResourcesHelper;
 export 'src/simdart.dart' hide SimDartHelper;
 export 'src/simulation_track.dart';
 export 'src/start_time_handling.dart';