+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Setting to determine how often `Future.delayed` is used instead of `Future.microtask` during event execution to allow GUI refresh.
 * Adding numeric properties and counter.
+* Allow creating new resources during simulation.
 
 ## 0.1.0